+++ src/Extractors/Csv.php
@@ -4,6 +4,7 @@ namespace Gettext\Extractors;
 
 use Gettext\Translations;
 use Gettext\Utils\HeadersExtractorTrait;
+use Gettext\Utils\CsvTrait;
 
 /**
  * Class to get gettext strings from csv.
@@ -11,6 +12,7 @@ use Gettext\Utils\HeadersExtractorTrait;
 class Csv extends Extractor implements ExtractorInterface
 {
     use HeadersExtractorTrait;
+    use CsvTrait;
 
     public static $options = [
         'delimiter' => ",",
@@ -29,7 +31,7 @@ class Csv extends Extractor implements ExtractorInterface
         fputs($handle, $string);
         rewind($handle);
 
-        while ($row = self::getcsv($handle, $options)) {
+        while ($row = self::fgetcsv($handle, $options)) {
             $context = array_shift($row);
             $original = array_shift($row);
 
@@ -47,20 +49,5 @@ class Csv extends Extractor implements ExtractorInterface
         }
 
         fclose($handle);
-    }
-
-    /**
-     * @param resource $handle
-     * @param array $options
-     *
-     * @return array
-     */
-    private static function getcsv($handle, $options)
-    {
-        if (version_compare(PHP_VERSION, '5.3.0') >= 0) { // >= 5.3
-            return fgetcsv($handle, null, $options['delimiter'], $options['enclosure'], $options['escape_char']);
-        }
-
-        return fgetcsv($handle, null, $options['delimiter'], $options['enclosure']);
     }
 }